refactor(pops-api): split backfill-test-fixtures under 200-line cap
The new WATCH_HISTORY_TABLE_SQL constant added in PRD-168 US-01 pushed
apps/pops-api/src/db/backfill-test-fixtures.ts to 213 non-blank,
non-comment lines, breaking the repo-wide oxlint max-lines (200) rule
that fails the PR #2999 Lint job.

Mirror the PR #2987 pattern of extracting siblings: split the SQL DDL
constants by pillar into

- backfill-test-fixtures-core.ts (service_accounts)
- backfill-test-fixtures-media.ts (shelf_impressions, movies, tv_shows,
  watch_history)
- backfill-test-fixtures-finance.ts (wish_list, entities, transactions,
  transaction_corrections, transaction_tag_rules, tag_vocabulary,
  budgets)

The original backfill-test-fixtures.ts becomes a barrel that re-exports
the same symbols so the three backfill-*-from-shared.test.ts consumers
keep their existing import surface untouched.

Line counts: core 22, media 93, finance 122, barrel 30 — all comfortably
under the cap. Full @pops/api and @pops/media-db vitest suites pass
unchanged; oxlint reports 0 errors.
